+++ src/components/category/CategoryDrawer.js
@@ -1,4 +1,4 @@
-import React, {useState} from 'react'
+import React, {useState, useEffect, useRef} from 'react'
 import Drawer from '@material-ui/core/Drawer'
 import List from '@material-ui/core/List'
 import ListItem from '@material-ui/core/ListItem'
@@ -7,92 +7,292 @@ import AddCircleOutlineIcon from '@material-ui/icons/AddCircleOutline'
 import DeleteIcon from '@material-ui/icons/Delete'
 import Paper from '@material-ui/core/Paper'
 import Input from '@material-ui/core/Input'
+import {AlertModal} from '../../components/modal'
+import {useCategoryState, useCategoryDispatch} from '../../containers/category/CategoryContainer'
 import CategoryHistoryContainer from '../../containers/category/CategoryHistoryContainer'
 import CategoryTab from './CategoryTab'
 import useStyles from './styles/CategoryDrawer'
 
 export default function CategoryDrawer(props) {
 
+  const categories = useCategoryState()
+  const dispatch = useCategoryDispatch()
   /*
   dispatch.getCategory()
   dispatch.writeCategory(value,1,false)
   이런식으로 함수 4가지 중에 하나 불러와서 사용 가능
   */
 
-  const { 
-    defaultCategories,
-    favoriteCategories,
-    children,
-  } = props
-
-  console.log(defaultCategories, favoriteCategories, children, )
+  
+  const favoritedArr = categories.filter(data => data.is_favorited === true)
+  const notFavoritedArr = categories.filter(data => data.is_favorited === false)
+  
+  console.log(favoritedArr, notFavoritedArr)
   
   const classes = useStyles()
   const [value, setValue] = useState('')
   const [selectedId, setSelectedId] = useState('')
-
+  const [addOpen, setAddOpen] = useState(true)
+  const [deleteOpen, setDeleteOpen] = useState(false)
+  const [deleteModalOpen, setDeleteModalOpen] = useState(false)
+  const [enterOpen, setEnterOpen] = useState(false)
 
   const handleChange = (e) => {
-    setValue(e.target.value);
+    setValue(e.target.value)
   }
-  const handleId = (id) => {
-    setSelectedId(id);
+
+  const addTab = () => {
+    dispatch.writeCategory(value, false)
+    setValue('')
+    setAddOpen(true)
+    setEnterOpen(false)
   }
+
+  const pressEnter = (e) => {
+    if (e.keyCode === 13) {
+      dispatch.writeCategory(value, false)
+      setValue('')
+      setAddOpen(true)
+      setEnterOpen(false)
+    }
+  }
+
+  const cancleAddTab = () => {
+    setAddOpen(true)
+    setEnterOpen(false)
+  }
+
+  const openDeleteModal = (e) => {
+    setDeleteOpen(true)
+    setDeleteModalOpen(true)
+  }
+
+  const closeDeleteModal = () => {
+    setDeleteModalOpen(false)
+    setDeleteOpen(false)
+    setAddOpen(true)
+  }
+  
+  const deleteTab = () => {
+    dispatch.deleteCategory(selectedId)
+    setDeleteModalOpen(false)
+    setDeleteOpen(false)
+    setAddOpen(true)
+  }
+  const toggleAddBtn = (id) => {
+    setAddOpen(false)
+    setDeleteOpen(true)
+    setSelectedId(id)
+  }
+
+  const toggleEnterTab = () => {
+    setAddOpen(false)
+    setEnterOpen(true)
+  }
+
+
+  /* 아래는 drag n drop 로직 */
+
+  const [dragged, setDragged] = useState('')
+  const [draggedOrder, setDraggedOrder] = useState(0)
+  const [draggedId, setDraggedId] = useState(0)
+  const [overedTabOrder, setOveredTabOrder] = useState(0)
+  const [overedTabFavorite, setOveredTabFavorite] = useState(null)
+  const [dragFinished, setDragFinished] = useState(false)
+
+  
+  const listRef = useRef()  
+
+  const dragStart = (e, id, order) => {
+    const target = e.currentTarget
+    setDragged(target)
+    setDraggedId(id)
+    setDraggedOrder(order)
+    e.dataTransfer.setData('text/html', target)
+  }
+
+  const dragOver = (e, order, favorited) => {
+    e.preventDefault()
+    setOveredTabOrder(order)
+    setOveredTabFavorite(favorited)
+    dragged.style.display='none'
+    e.currentTarget.previousSibling.style.display = 'block'
+
+    console.log('over',order, favorited)
+  }
+
+  const dragLeave = (e) => {
+    e.currentTarget.previousSibling.style.display = 'none'
+  }
+  
+  const dragEnd = (e, id, name, order, favorited) => {
+
+    dispatch.updateCategory(id, name, order, favorited)
+    setDragFinished(true)
+    document.querySelectorAll('.dragline').forEach(el => {
+      el.style.display = 'none'
+    })
+    console.log('end', name, favorited)
+  }
+
+  const firstFavoriteDragOver = (e) => {
+    e.preventDefault()
+    dragged.style.display='none'
+    setOveredTabOrder(draggedOrder)
+    setOveredTabFavorite(true)
+  }
+
+
+  /* 아래는 외부영역 클릭시 버튼 토글 & cleartimeout */
+
+  const wrapperRef = useRef(null)
+  const timeId = useRef()
+
+  useEffect(() => {
+
+    //change add&delete button state if clicked on outside of element
+    function handleClickOutside(event) {
+      if (wrapperRef.current && !wrapperRef.current.contains(event.target)) {
+        setAddOpen(true)
+        setDeleteOpen(false)
+        console.log(wrapperRef.current, event.target)
+      } 
+    }
+    // Bind the event listener
+    document.addEventListener("mousedown", handleClickOutside);
+
+    //setTimeout
+    if(dragFinished) {
+      timeId.current = setTimeout(() => setDragFinished(false), 800)
+    }
+
+    return () => {
+      // Unbind the event and timeout on clean up
+      document.removeEventListener("mousedown", handleClickOutside)
+      //clearTimeout
+      clearTimeout(timeId.current)
+    }
+
+  },[wrapperRef, dragFinished])
+
+
   
   const drawer = (
     <div>
-      <div className="list-tab-layout">
+      <div className="list-tab-layout" ref={wrapperRef}>
         <div className="favorite-text">
           Favorite
         </div>
         <hr />
-        <div className="drag-box">
+        <div 
+          className={(favoritedArr.length === 0 ? 'drag-box' : 'hidden')}
+          onDragOver={firstFavoriteDragOver}
+        >
           Drag the category here!
         </div>
-        <List>
-          {favoriteCategories.map((data, index) => (
-            <ListItem key={index} 
+        <List 
+        ref={listRef}>
+          {favoritedArr.map((data, index) => (
+            <>
+            <div className={classes.dragline + " dragline"}></div>
+            <ListItem 
+              key={data.id} 
               className={classes.listItem + (data.id === selectedId ? ' '+classes.selected : '' )}
-              onClick={() => handleId(data.id)}>
-              <CategoryTab key={index} text={data.name} id={data.id}/>
+              onClick={() => toggleAddBtn(data.id)}
+              draggable='true'
+              onDragStart={(e) => dragStart(e, data.id, data.order)}
+              onDragEnd={(e) => dragEnd(e, data.id, data.name, overedTabOrder, overedTabFavorite)}
+              onDragOver={(e) => dragOver(e, (draggedOrder < data.order ? data.order-1 : data.order) , data.is_favorited)}
+              onDragLeave={dragLeave}
+            >
+              <CategoryTab 
+                key={data.id} 
+                text={data.name} 
+                id={data.id} 
+                order={data.order}
+                isFavorited={data.is_favorited}
+                urlCount={data.url_count}
+                selected={(data.id === selectedId)} 
+                dragFinished={(data.id === draggedId ? dragFinished : false)} 
+              />
             </ListItem>
+            </>
           ))}
         </List>
         <div className="category-text">
           Category
         </div>
         <hr />
-        <Button className={classes.tabButton} variant="contained">
+        <Button 
+          className={classes.addButton + (addOpen ? '' : ' '+classes.hidden)} 
+          variant="contained"
+          onClick={toggleEnterTab}
+        >
           <AddCircleOutlineIcon style={{color: "#cccccc"}} />
         </Button>
-        <Button className={classes.tabButton} variant="contained">
+        <Button 
+          className={classes.deleteButton + (deleteOpen ? ' '+classes.block : '')} 
+          variant="contained" 
+          onClick={openDeleteModal}
+        >
           <DeleteIcon style={{color: "#cccccc"}} />
         </Button>
-        <Paper component="div" className={classes.enterTab}>
+        <Paper 
+          component="div" 
+          className={classes.enterTab + (enterOpen ? ' '+classes.flex : '')}
+        >
           <Input
             disableUnderline={true}
             className={classes.input}
             placeholder="New one"
             value={value}
             onChange={handleChange}
+            onKeyDown={pressEnter}
           />
-            <Button className={classes.okBtn}>확인</Button>
-            <Button className={classes.cancleBtn}>취소</Button>
+            <Button className={classes.okBtn} onClick={addTab}>확인</Button>
+            <Button className={classes.cancleBtn} onClick={cancleAddTab}>취소</Button>
         </Paper>
         <List>
-        {defaultCategories.map((data, index) => (
-          <ListItem key={index} 
+        {notFavoritedArr.map((data, index) => (
+          <>
+          <div className={classes.dragline + " dragline"}></div>
+          <ListItem 
+            key={data.id} 
             className={classes.listItem + (data.id === selectedId ? ' '+classes.selected : '' )}
-            onClick={() => handleId(data.id)}>
-            <CategoryTab key={index} text={data.name} id={data.id} />
-          </ListItem>
+            onClick={() => toggleAddBtn(data.id)}
+            draggable='true'
+            onDragStart={(e) => dragStart(e, data.id, data.order)}
+            onDragEnd={(e) => dragEnd(e, data.id, data.name, overedTabOrder, overedTabFavorite)}
+            onDragOver={(e) => dragOver(e, (draggedOrder < data.order ? data.order-1 : data.order) , data.is_favorited)}
+            onDragLeave={dragLeave}
+          >
+            <CategoryTab 
+            key={data.id} 
+            text={data.name} 
+            id={data.id} 
+            order={data.order}
+            isFavorited={data.is_favorited}
+            urlCount={data.url_count}
+            selected={(data.id === selectedId)} 
+            dragFinished={(data.id === draggedId ? dragFinished : null)} 
+            />
+            </ListItem>
+          </>
         ))}
-        </List>   
+        </List>
       </div>
+
+      <AlertModal 
+        btnText = '삭제'
+        modalText = '카테고리를 삭제하면 안에 저장된 모든 탭이 삭제 됩니다. 그래도 삭제 하시겠습니까?'
+        openBool = {deleteModalOpen} 
+        onClose = {closeDeleteModal}
+        onClickOk = {deleteTab} />
+
     </div>
   )
 
-  console.log('appBar', props.appBar, props)
+
 
   return (
     <div className={classes.root}>
@@ -109,7 +309,7 @@ export default function CategoryDrawer(props) {
       </nav>
       <main className={classes.content}>
         <div className={classes.toolbar} />
-        {children}
+        {props.children}
       </main>
       <CategoryHistoryContainer />
     </div>
