+++ src/components/category/CategoryTab.js
@@ -1,13 +1,14 @@
-import React, {useState} from 'react'
-import Paper from '@material-ui/core/Paper'
-import Input from '@material-ui/core/Input'
+import React, {useState} from 'react';
+import Paper from '@material-ui/core/Paper';
+import InputBase from '@material-ui/core/Input';
+import {useCategoryDispatch} from '../../containers/category/CategoryContainer';
 import useStyles from './styles/CategoryTab'
-// import {useCategoryDispatch} from '../../containers/CategoryContainer';
 
 
-export default function CategoryTab({text, id}) {
+export default function CategoryTab({text, id, order, isFavorited, urlCount, selected, dragFinished}) {
   const classes = useStyles()
-  // const dispatch = useCategoryDispatch()
+
+  const dispatch = useCategoryDispatch()
   const [value, setValue] = useState(text);
   const [disabled, setDisabled] = useState(true);
 
@@ -19,30 +20,30 @@ export default function CategoryTab({text, id}) {
     setDisabled(!disabled)
   }
 
-  // const updateText = (e) => {
-  //   console.log(id)
-  //     if (e.keyCode === 13) {
-  //       dispatch.updateCategory(id, value, false )
-  //       setDisabled(!disabled)
-
-  //     }
-  // }
+  const updateText = (e) => {
+      if (e.keyCode === 13) {
+        dispatch.updateCategory(id, value, order, isFavorited )
+        setDisabled(!disabled)
+      }
+  }
 
   return (
     <div>
       <Paper 
         component="div" 
-        className={classes.root} 
+        className={classes.root + (dragFinished ? ' dragFinished' : '') } 
+        id={`${id}`}
       >
-        <Input
+        <InputBase 
           disableUnderline={true}
-          className={classes.input}
+          className={classes.input + (selected ? ' selected': '')}
           disabled={disabled}
           onDoubleClick={onDoubleClick}
           value={value}
           onChange={handleChange}
-          // onKeyDown={updateText}
+          onKeyDown={updateText}
         />
+        <div className={classes.urlCountBox}>탭 {urlCount}개</div>
       </Paper>
     </div>
   )
