+++ src/components/category/styles/CategoryTab.js
@@ -9,7 +9,22 @@ const useStyles = makeStyles((theme) => ({
     padding:12,
     fontFamily: "SpoqaHanSans",
     fontSize: 16,
+    width: "70%"
+  },
+  selected : {
+    fontSize: 16,
+    fontWeight: "bold",
+    color: '#2083ff'
+  },
+  urlCountBox : {
+    width: "30%",
+    display: "inline-block",
+    fontSize: 12,
+    fontFamily: "SpoqaHanSans",
+    color: "#868e96",
+    textAlign: "center"
   }
+
 }))
 
 export default useStyles
